fix: added proper naming convention
