+++ package.json
@@ -15,6 +15,7 @@
     "@iarna/toml": "^2.2.5",
     "@rover-tools/engine": "file:../rover-engine",
     "@types/inquirer": "^8.2.1",
+    "dotenv": "^16.0.3",
     "eslint-config-prettier": "^8.5.0",
     "esm": "^3.2.25",
     "inquirer": "^8.2.4",
@@ -34,10 +35,10 @@
     "start": "tsc   bin/init.ts && tsc  bin/index.ts && npm install -g . ",
     "build": "tsc && npm run format:write && npm run lint:fix",
     "watch": "tsc --watch",
-    "localtest": "npm install rover-engine && npm run build && npm install rover-cli -g  ",
+    "localtest": "npm install $(grep rover_engine .env | cut -d '=' -f2 | tr -d '\"' ) && npm run build && npm install $(grep rover_cli .env | cut -d '=' -f2 | tr -d '\"') -g  ",
     "prepack": "npm run clean && npm run build ",
     "format:check": "prettier --check .",
-    "format:write": "prettier --write .",
+    "format:write": "prettier --write src/* bin/*",
     "lint:check": "eslint . ",
     "lint:fix": "eslint --fix ."
   },
